Finishing Ice Level with no erros

+++ ProjectGame/Project Game/Assets/Scripts/Active Scripts/Scripts - Audra/IPickUp.cs
@@ -12,5 +12,5 @@ public interface IPickUp
     {
 
     }
-    public bool gainFuel(int fuel);
+    public bool gainFuel(float fuel);
 }
